requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.10.6
PyJWT==2.10.1
tortoise-orm==0.24.1
aiomysql==0.2.0